+++ test/provider/events.js
@@ -255,11 +255,26 @@ describe("Provider events", () => {
       expect(spy.callCount).toEqual(1);
     });
 
+    it("should emit cleanCache options as first argument", async () => {
+      const spy = sandbox.spy();
+      provider.on("cleanCache", spy);
+      provider.cleanCache({ force: true });
+      expect(spy.getCall(0).args[0]).toEqual({ force: true });
+    });
+
     it("should emit a cleanCache event when listener is added with wildcard", async () => {
       const spy = sandbox.spy();
       provider.on("*", spy);
       provider.cleanCache();
       expect(spy.getCall(0).args[0]).toEqual("cleanCache");
+    });
+
+    it("should emit cleanCache options as second argument when listener is added with wildcard", async () => {
+      const spy = sandbox.spy();
+      provider.on("*", spy);
+      provider.cleanCache({ force: true });
+      expect(spy.getCall(0).args[0]).toEqual("cleanCache");
+      expect(spy.getCall(0).args[1]).toEqual({ force: true });
     });
 
     it("should emit a child cleanCache event when child cache is clean", async () => {
@@ -269,11 +284,27 @@ describe("Provider events", () => {
       expect(spy.callCount).toEqual(1);
     });
 
+    it("should emit cleanCache options as second argument when child cache is clean", async () => {
+      const spy = sandbox.spy();
+      provider.onChild("cleanCache", spy);
+      childProvider.cleanCache({ force: true });
+      expect(spy.getCall(0).args[0]).toEqual(childProvider);
+      expect(spy.getCall(0).args[1]).toEqual({ force: true });
+    });
+
     it("should emit a child cleanCache event when parent cache is clean", async () => {
       const spy = sandbox.spy();
       provider.onChild("cleanCache", spy);
       provider.cleanCache();
       expect(spy.callCount).toEqual(1);
+    });
+
+    it("should emit cleanCache options as second argument when parent cache is clean", async () => {
+      const spy = sandbox.spy();
+      provider.onChild("cleanCache", spy);
+      provider.cleanCache({ force: true });
+      expect(spy.getCall(0).args[0]).toEqual(childProvider);
+      expect(spy.getCall(0).args[1]).toEqual({ force: true });
     });
   });