+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - chore(ci): Migrate from Travis CI to github actions
 - chore(deps): Support all Node.js releases that have not passed their end date
-
 ### Changed
+- test(mutation): Set branch name in stryker dashboard based on environment variable
 ### Fixed
+- fix(events): Pass force option also to children when cleaning caches
+- fix(selector): Force cleaning of cache when a selector dependency is cleaned
+- fix(providers): Allow options in providers.cleanCache method
 ### Removed
 
 ## [2.8.1] - 2020-11-27